requirements.txt
numpy
faiss-cpu
insightface
onnxruntime
opencv-python
streamlit
albumentations
